LibJS+LibTest: Move out the test-js test runner into LibTest
