+++ Meta/Lagom/CMakeLists.txt
@@ -132,7 +132,9 @@ if (BUILD_LAGOM)
         set_target_properties(ntpquery_lagom PROPERTIES OUTPUT_NAME ntpquery)
         target_link_libraries(ntpquery_lagom Lagom)
 
-        add_executable(test-js_lagom ../../Tests/LibJS/test-js.cpp)
+        add_executable(test-js_lagom
+            ../../Tests/LibJS/test-js.cpp
+            ../../Userland/Libraries/LibTest/JavaScriptTestRunnerMain.cpp)
         set_target_properties(test-js_lagom PROPERTIES OUTPUT_NAME test-js)
         target_link_libraries(test-js_lagom Lagom)
         target_link_libraries(test-js_lagom stdc++)
